Separate resource preparation from build script
Cherry-picked from https://github.com/GooborgStudios/mdn-bcd-collector/commit/1577d8a6e865ed8b8083f0f3c02b0b0787560e02.
