+++ package.json
@@ -18,13 +18,14 @@
     "format:fix": "eslint --fix . && prettier --write \"**/*.json\" \"**/*.md\" \"**/*.scss\" \"**/*.yaml\"",
     "fix": "npm run format:fix",
     "clean": "rm -rf .nyc_output coverage generated tests.json",
-    "build": "npm run build:compile && npm run build:tests",
+    "build": "npm run build:compile && npm run build:tests && npm run build:resources",
     "build:compile": "tsc -p .",
     "build:tests": "ts-node build.ts",
+    "build:resources": "ts-node prepare-resources.ts",
     "gcp-build": "npm run build",
     "deploy": "gcloud app deploy",
     "start": "NODE_ENV=production node app.js",
-    "dev": "npm i && nodemon -e ts,json,idl,scss,yaml --exec \"npm run build:tests && ts-node app.ts\"",
+    "dev": "npm i && nodemon -e ts,json,idl,scss,yaml --exec \"npm run build:tests && npm run build:resources && ts-node app.ts\"",
     "unittest": "ts-node scripts.ts unittest",
     "coverage": "c8 report -r lcov && open-cli coverage/lcov-report/index.html",
     "test": "npm run format && npm run unittest",
